First commit
Unrandom is a Python script to play videos in a pseudo-random fashion
based on their user-provided rating.

When it is run, it prompts the user to 1) build a SQL database of the
files in directory and subdirectories, 2) rate the files in the
database, or 3) play a randomized file.

The ratings range from 0-3: 0 is never selected, 1 is selected 1/21
times, 2 is selected 4/21 times, and 3 is selected 16/21 times.

Requires a local SQL server to be running; built with MariaDB in mind.